+++ packages/angular/src/index.ts
@@ -19,6 +19,7 @@ export * from './scatter-chart.component';
 export * from './meter-chart.component';
 export * from './radar-chart.component';
 export * from './combo-chart.component';
+export * from './tree-chart.component';
 export * from './treemap-chart.component';
 export * from './circle-pack-chart.component';
 export * from './wordcloud-chart.component';