+++ packages/core/src/configuration.ts
@@ -17,6 +17,7 @@ import {
 	HistogramChartOptions,
 	RadarChartOptions,
 	ComboChartOptions,
+	TreeChartOptions,
 	TreemapChartOptions,
 	CirclePackChartOptions,
 	WorldCloudChartOptions,
@@ -43,6 +44,7 @@ import {
 	ToolbarControlTypes,
 	ZoomBarTypes,
 	LegendItemType,
+	TreeTypes,
 } from './interfaces';
 import enUSLocaleObject from 'date-fns/locale/en-US/index';
 import { circlePack } from './configuration-non-customizable';
@@ -542,6 +544,19 @@ const comboChart: ComboChartOptions = Tools.merge({}, baseBarChart, {
 } as ComboChartOptions);
 
 /*
+ * options specific to tree charts
+ */
+const treeChart: TreeChartOptions = Tools.merge(
+	{
+		tree: {
+			type: TreeTypes.TREE,
+		},
+	},
+	chart,
+	{} as TreeChartOptions
+);
+
+/*
  * options specific to treemap charts
  */
 const treemapChart: TreemapChartOptions = Tools.merge({}, chart, {
@@ -597,6 +612,7 @@ export const options = {
 	radarChart,
 	gaugeChart,
 	comboChart,
+	treeChart,
 	treemapChart,
 	circlePackChart,
 	wordCloudChart,